+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 * Emit colors using their original representation if possible.
 
+* Emit colors without an original representation as names if possible.
+
 ## 1.0.0-alpha.2
 
 * Fix a bug where variables, functions, and mixins were broken in imported